proyecto atualizado

+++ src/main/java/com/example/SistemaEnvio/service/ClientesService.java
@@ -2,10 +2,16 @@ package com.example.SistemaEnvio.service;
 
 
 import com.example.SistemaEnvio.entity.Clientes;
+import com.example.SistemaEnvio.exception.ResourceNotFoundException;
 import com.example.SistemaEnvio.repository.ClientesRepository;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Service
@@ -21,8 +27,11 @@ public class ClientesService {
         return clientesRepository.findAll();
     }
 
-    public Optional<Clientes> getClientesById(Long id){
-        return  clientesRepository.findById(id);
+    public ResponseEntity<Clientes> getClientesById(Long id){
+         Clientes clientes = clientesRepository.findById(id)
+                 .orElseThrow(() ->new ResourceNotFoundException("el cliente no existe con ese id: " + id));
+    return  ResponseEntity.ok(clientes);
+
     }
 
     public Clientes saveClientes(Clientes clientes){
@@ -30,8 +39,29 @@ public class ClientesService {
 
     }
 
-    public void deleteClientes(Long id){
+    public ResponseEntity<Clientes> updateClientes(Long id, @RequestBody Clientes clientes){
+        Clientes cliente = clientesRepository.findById(id)
+                .orElseThrow(()-> new ResourceNotFoundException("El Cliente no existe con ese id: "+id));
+
+        cliente.setIdentificacion(cliente.getIdentificacion());
+        cliente.setNombres(clientes.getNombres());
+        cliente.setApellidos(clientes.getApellidos());
+        cliente.setDireccion(clientes.getDireccion());
+        cliente.setTelefono(clientes.getTelefono());
+        cliente.setEmail(clientes.getEmail());
+
+        Clientes clienteActualizado =clientesRepository.save(cliente);
+        return ResponseEntity.ok(clienteActualizado);
+    }
+
+    public ResponseEntity<Map<String,Boolean>> deleteClientes(Long id){
+        if (!clientesRepository.existsById(id)) {
+             throw new ResourceNotFoundException("Cliente no existe!");
+        }
         clientesRepository.deleteById(id);
+        Map<String,Boolean> response = new HashMap<>();
+        response.put("Eliminado con exito",Boolean.TRUE);
+        return ResponseEntity.ok(response);
     }
 
 }
